chore(pkger): axe deprecated /packges route tree
